feat: 비즈니스프로필 카테고리 제거 및 리팩토링(#83)
