+++ backend/src/main/java/com/minibeit/businessprofile/domain/BusinessProfile.java
@@ -23,8 +23,6 @@ public class BusinessProfile extends BaseEntity {
 
     private String name;
 
-    private String category;
-
     private String place;
 
     private String introduce;
@@ -39,34 +37,26 @@ public class BusinessProfile extends BaseEntity {
     @OneToMany(mappedBy = "businessProfile", cascade = CascadeType.ALL)
     private List<UserBusinessProfile> userBusinessProfileList = new ArrayList<>();
 
-    private void addUser(UserBusinessProfile userBusinessProfile) {
-        this.userBusinessProfileList.add(userBusinessProfile);
-        userBusinessProfile.setBusinessProfile(this);
-    }
-
     public void update(BusinessProfileRequest.Update createRequest) {
         this.name = createRequest.getName();
-        this.category = createRequest.getCategory();
         this.place = createRequest.getPlace();
         this.contact = createRequest.getContact();
         this.introduce = createRequest.getIntroduce();
-    }
-
-    public static BusinessProfile create(BusinessProfileRequest.Create request, UserBusinessProfile userBusinessProfile, File avatar) {
-        BusinessProfile businessProfile = BusinessProfile.builder()
-                .name(request.getName())
-                .category(request.getCategory())
-                .place(request.getPlace())
-                .introduce(request.getIntroduce())
-                .contact(request.getContact())
-                .avatar(avatar)
-                .build();
-        businessProfile.addUser(userBusinessProfile);
-        return businessProfile;
     }
 
     public void updateAvatar(File avatar) {
         this.avatar = avatar;
     }
 
+    public static BusinessProfile create(BusinessProfileRequest.Create request, UserBusinessProfile userBusinessProfile, File avatar) {
+        BusinessProfile businessProfile = BusinessProfile.builder()
+                .name(request.getName())
+                .place(request.getPlace())
+                .introduce(request.getIntroduce())
+                .contact(request.getContact())
+                .avatar(avatar)
+                .build();
+        userBusinessProfile.setBusinessProfile(businessProfile);
+        return businessProfile;
+    }
 }
