+++ backend/src/main/java/com/minibeit/businessprofile/dto/BusinessProfileResponse.java
@@ -29,7 +29,6 @@ public class BusinessProfileResponse {
     public static class GetOne {
         private Long id;
         private String name;
-        private String category;
         private String place;
         private String introduce;
         private String contact;
@@ -40,7 +39,6 @@ public class BusinessProfileResponse {
             GetOneBuilder getOneBuilder = GetOne.builder()
                     .id(businessProfile.getId())
                     .name(businessProfile.getName())
-                    .category(businessProfile.getCategory())
                     .place(businessProfile.getPlace())
                     .introduce(businessProfile.getIntroduce())
                     .contact(businessProfile.getContact())
