+++ backend/src/main/java/com/minibeit/businessprofile/service/BusinessProfileService.java
@@ -86,11 +86,9 @@ public class BusinessProfileService {
         if (userBusinessProfileRepository.existsByUserIdAndBusinessProfileId(userToShare.getId(), businessProfileId)) {
             throw new DuplicateShareException();
         }
-        UserBusinessProfile userBusinessProfile = UserBusinessProfile.create(userToShare);
-        userBusinessProfile.createWithBusinessProfile(businessProfile);
+        UserBusinessProfile userBusinessProfile = UserBusinessProfile.createWithBusinessProfile(userToShare, businessProfile);
 
         userBusinessProfileRepository.save(userBusinessProfile);
-
     }
 
     public void cancelShare(Long businessProfileId, Long userId, User user) {
